Improved and organised code for dialogue box patch Moved letter viewer menu patch to its own class moved npc patch to its own clas moved draw hover text patch to IClickableMenuPatch.cs
